+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Next release
+
+* Address upcoming [Trakt pagination requirements](https://github.com/trakt/trakt-api/discussions/775) for watched methods:
+  * Add paginated variants of `Sync.watchedMovies()`, `Sync.watchedShows()`, `Users.watchedMovies()` and 
+    `Users.watchedShows()` with required `page` and `limit` parameters. The original methods are deprecated.
+* Add `min` variants of watched methods: `Users.watchedMoviesMin()`, `Users.watchedShowsMin()`, `Sync.watchedMoviesMin()` and `Sync.watchedShowsMin()`
+* Add episode variant of watched methods: `Sync.watchedEpisodes()` and `Sync.watchedEpisodesMin()` 
+
 ## 6.18.0 - 2026-03-06
 
 * Watchlist endpoints require pagination: deprecate the variants that don't accept page and limit parameters.